requirements.txt
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.21.0
httplib2==0.20.1
oauthlib==3.1.1
requests==2.26.0
requests-oauthlib==1.3.0
urllib3==1.26.7
flask==2.0.2
flask-cors==3.0.10
flask-restful==0.3.9
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.25
pytz==2021.3
python-dateutil==2.8.2
pandas==1.3.4
numpy==1.21.2
pytest==6.2.5
pytest-cov==3.0.0
coverage==6.1.2
mock==4.0.3
requests-mock==1.9.3
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
python-dotenv==0.19.1
python-jose==3.3.0
bcrypt==3.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-decouple==3.5
